plugins: psbt_nostr: split generic and UI parts
